prova_langchain_playwright/pyproject.toml
[project]
name = "prova-langchain-playwright"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.21",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]
